move cmplx npy_format_descriptor template to array_common.hpp
