+++ cpp/modmesh/math/CMakeLists.txt
@@ -7,7 +7,6 @@ set(MODMESH_MATH_HEADERS
 
 set(MODMESH_MATH_PYMODHEADERS
     ${CMAKE_CURRENT_SOURCE_DIR}/pymod/math_pymod.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/pymod/wrap_Complex.hpp
     CACHE FILEPATH "" FORCE)
 
 set(MODMESH_MATH_PYMODSOURCES
